Create README.md
This project  analyzes stock market data between 2014 to 2016 with VBA. During the analysis, two data-points are used. Test Data is used while developing the script, and Stock Data is the main data to run the script.  Sine this file is bigger in size, it may take time when you execute the script into the file. 

Code description
This repository contains the VBA code in the VBA script file. This script can be run in the .xlsm files as macros to get the corresponding results.